Move chat code into module. Fix pebcak.
