+++ client/chat.ts
@@ -1,18 +1,23 @@
-import { sendChatMessage } from './server';
+import { sendChatMessage } from './server.js';
 
+// Represents a chat message
 interface Message {
     id: string;
     body: string;
 }
 
+// If we need to pre-populate the chat with previous messages, they get
+// loaded from here.
 const chatData: Message[] = [];
 
 /**
- * Take in a `Message` object and return an HTML string.
+ * Take in a `Message` object and return an HTML element with the chat
+ * message.
  */
 const formatMessage = (data: Message): HTMLElement => {
     const div = document.createElement('div');
     div.classList.add('message');
+    // TODO: the username should eventually be better than a sliced random ID.
     div.innerHTML = `
         <span class="username">${data.id.slice(
             -5
@@ -22,12 +27,35 @@ const formatMessage = (data: Message): HTMLElement => {
 };
 
 /**
- * Take in a message string and append to the DOM.
+ * Take in a message's HTML element and append it to the DOM.
  */
 const printMessage = (message): void => {
     const messageOutputArea = document.getElementById('messages');
     messageOutputArea.appendChild(message);
 };
 
+/**
+ * Send new messages
+ */
+const initializeChatListener = (socket): void => {
+    const chatForm: HTMLElement = document.getElementById('chatForm');
 
-export { chatData, formatMessage, printMessage };
+    chatForm.addEventListener('submit', (e): void => {
+        e.preventDefault();
+        const chatMessageInput = <HTMLInputElement>(
+            document.getElementById('chatMessageInput')
+        );
+        const chatMessage: string = chatMessageInput.value.trim();
+        chatMessageInput.value = '';
+        console.log('from form:', chatMessage);
+        sendChatMessage(socket, chatMessage);
+    });
+
+    socket.on('chat', (message): void => {
+        console.log('from server:', message);
+        const messageHtml = formatMessage(message);
+        printMessage(messageHtml);
+    });
+};
+
+export { chatData, formatMessage, printMessage, initializeChatListener };
