+++ templates/index.html
@@ -20,9 +20,8 @@
         <link rel="stylesheet" href="/static/styles/main.css" />
     </head>
     <body>
-        <noscript
-            >This game requires JavaScript. Please turn on JavaScript.</noscript
-        >
+        <noscript>Please enable JavaScript to play this game.</noscript>
+
         <canvas id="canvas" width="480" height="320"></canvas>
 
         <!-- Chat could be moved into canvas later, if needed. -->
